Enhance BTC fee service to provide real time BTC pricing for TX
Provides an enhanced data structure in RMQ and Redis which contains:

{
  "recFeeInSatPerByte":160,
  "recFeeInSatForAvgTx":37600,
  "recFeeInBtcForAvgTx":0.000376,
  "recFeeInUsdForAvgTx":0.48,
  "avgTxSizeBytes":235
}

This provides some data about what the actual USD cost for publishing
a TX will be based on the current cost measured in satoshis/byte multiplied
by the average tx size and converts it to btc/tx and usd/tx.

There is a primary and a fallback provider of exchange data. If both
are unavailable the services continues to work, just drops the USD
conversion data and returns null for that value instead.